chore: Refactor file structure to align with FCIS Vertical Slice Architecture
