+++ src/features/read/mod.rs
@@ -1,2 +1,5 @@
-pub mod read;
-pub mod tests_read;
+pub mod handler;
+pub mod logic;
+pub mod model;
+mod tests;
+
