+++ src/shared/core/adapters/mod.rs
