+++ src/shared/core/errors/file_io.rs
