+++ src/shared/core/errors/mod.rs
