+++ src/shared/core/errors/not_supported.rs
