+++ src/shared/core/errors/path.rs
