+++ src/shared/core/errors/value_type.rs
