+++ src/shared/core/mod.rs
@@ -1,0 +1,6 @@
+pub mod adapters;
+pub mod errors;
+pub mod parse;
+pub mod path;
+pub mod suggest;
+pub mod types;
