+++ src/shared/core/parse.rs
@@ -1,0 +1,38 @@
+use core::fmt;
+use std::path::PathBuf;
+
+use crate::shared::core::errors::ParseError;
+
+// ----- PARSE TYPES -----
+#[derive(Debug, Clone, PartialEq)]
+pub struct SourceLocation {
+    pub line: usize,
+    pub column: usize,
+    pub file: Option<PathBuf>,
+}
+
+impl SourceLocation {
+    pub fn new(line: usize, column: usize) -> Self {
+        SourceLocation {
+            line,
+            column,
+            file: None,
+        }
+    }
+
+    pub fn with_file(mut self, file: PathBuf) -> Self {
+        self.file = Some(file);
+        self
+    }
+}
+
+impl fmt::Display for SourceLocation {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match &self.file {
+            Some(p) => write!(f, "{}:{}:{}", p.display(), self.line, self.column),
+            None => write!(f, "{}:{}", self.line, self.column),
+        }
+    }
+}
+
+pub type ParseResult<T> = Result<T, ParseError>;
