+++ src/shared/mod.rs
@@ -1,4 +1,2 @@
-pub mod adapters;
-pub mod errors;
-pub mod helpers;
-pub mod types;
+pub mod core;
+pub mod shell;
